[ refactor: issue-37 ] rename shared/variables to shared/constants; cleanup of shared/constants
