+++ libs/shared/constants/.eslintrc.json
