+++ libs/shared/constants/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@mentorship/shared/variables",
+  "name": "@mentorship/shared/constants",
   "version": "0.0.1",
   "type": "commonjs"
 }
