+++ libs/shared/constants/project.json
@@ -1,0 +1,26 @@
+{
+  "name": "shared-constants",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/shared/constants/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/libs/shared/constants",
+        "main": "libs/shared/constants/src/index.ts",
+        "tsConfig": "libs/shared/constants/tsconfig.lib.json",
+        "assets": ["libs/shared/constants/*.md"]
+      }
+    },
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["libs/shared/constants/**/*.ts"]
+      }
+    }
+  },
+  "tags": []
+}
