+++ libs/shared/constants/src/index.ts
@@ -1,0 +1,3 @@
+export { font } from './font';
+export { space } from './spacing';
+export { Color } from './color';
