+++ libs/shared/constants/tsconfig.json
@@ -14,9 +14,6 @@
   "references": [
     {
       "path": "./tsconfig.lib.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
     }
   ]
 }
