+++ libs/shared/constants/tsconfig.lib.json
