+++ libs/shared/ui/src/components/avatar-modal/AvatarModal.styled.tsx
@@ -1,4 +1,4 @@
-import { Color } from '@mentorship/shared/variables';
+import { Color } from '@mentorship/shared/constants';
 import styled from 'styled-components';
 
 import { ImageBody } from '../image/Image.styled';
