+++ libs/shared/ui/src/components/avatar/ChangeAvatar.styled.tsx
@@ -1,4 +1,4 @@
-import { Color, space } from '@mentorship/shared/variables';
+import { Color, space } from '@mentorship/shared/constants';
 import styled from 'styled-components';
 
 import { ButtonBaseStyles } from '../button/Button.styled';
