+++ libs/shared/ui/src/components/loader/Loader.styled.tsx
@@ -1,4 +1,4 @@
-import { Color } from '@mentorship/shared/variables';
+import { Color } from '@mentorship/shared/constants';
 import styled from 'styled-components';
 
 export const SpinnerWrapper = styled.div`
