--- libs/shared/variables/README.md
@@ -1,11 +1,0 @@
-# shared-variables
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build shared-variables` to build the library.
-
-## Running unit tests
-
-Run `nx test shared-variables` to execute the unit tests via [Jest](https://jestjs.io).
