--- libs/shared/variables/src/index.ts
@@ -1,3 +1,0 @@
-export { font } from './font';
-export { space } from './spacing';
-export { Color } from './color';
