--- libs/shared/variables/src/spacing.ts
@@ -1,12 +1,0 @@
-const BASE_SPACE = 4;
-
-export const space = {
-  x2: `${2 * BASE_SPACE}px`,
-  x3: `${3 * BASE_SPACE}px`,
-  x4: `${4 * BASE_SPACE}px`,
-  x5: `${5 * BASE_SPACE}px`,
-  x6: `${6 * BASE_SPACE}px`,
-  x7: `${7 * BASE_SPACE}px`,
-  x8: `${8 * BASE_SPACE}px`,
-  x20: `${20 * BASE_SPACE}px`,
-} as const;
